rafactor to gradle
</subject>

Branch: dev

<type>:
- [ ] Bug fix
- [ ] Bug fix (Test)
- [x] New feature
- [x] Breaking change
- [ ] Documentation update
- [x] This change requires a documentation update

<body>

<footer>
