+++ .gitignore
@@ -224,3 +224,5 @@ cython_debug/
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 
+gradle/
+.gradle/
