+++ include/Redirect.java
@@ -1,5 +1,3 @@
-package include;
-
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
