+++ lab00/src/lab_00_C.java
@@ -1,6 +1,3 @@
-package lab_00;
-
-
 import java.io.DataInputStream;
 import java.io.FileInputStream;
 import java.io.IOException;
